the-luminal-archive/quantum-memory/requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# CUDA and quantum libraries
cupy-cuda12x
cuquantum-python>=24.03.0
qiskit>=0.45.0
pennylane>=0.33.0

# Emotion processing
transformers>=4.36.0
torch>=2.0.0
datasets>=2.14.0

# Memory and storage
h5py>=3.9.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Web and API
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.5.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0
psutil>=5.9.0